toas not i izmena i brisanje

+++ MyProjectBackendSolution/MyProjectBackend/Services/ThemeService.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using Microsoft.EntityFrameworkCore;
 using MyProjectBackend.Dto.Theme;
 using MyProjectBackend.Infrastructure;
 using MyProjectBackend.Interfaces;
@@ -27,7 +28,28 @@ namespace MyProjectBackend.Services
 
         public bool DeleteTheme(int id)
         {
-            
+            // Temu mozemo obrisati samo ako ne postoji uradjeni kviz sa tom temom
+            var quizzes = _dbContext.UserQuizzs
+               .Select(uq => uq.QuizzId)
+               .Distinct()
+               .ToList();
+
+            var usedQuizzes = _dbContext.Quizzes
+                .Where(q => quizzes.Contains(q.Id))
+                .Include(q => q.Themes) 
+                .ToList();
+
+           
+            bool existsInQuizzes = usedQuizzes
+                .Any(q => q.Themes.Any(qt => qt.Id == id));
+
+            bool hasQuestion = _dbContext.Questions.Where(q=>q.ThemeId == id).Any();
+
+            if (existsInQuizzes || hasQuestion)
+            {
+                return false; // ne dozvoli brisanje
+            }
+
             Theme theme = _dbContext.Themes.Find(id);
 
             _dbContext.Themes.Remove(theme);
@@ -50,6 +72,24 @@ namespace MyProjectBackend.Services
 
         public ThemeDto UpdateTheme(int id, ThemeDto updatedTheme)
         {
+            var quizzes = _dbContext.UserQuizzs
+              .Select(uq => uq.QuizzId)
+              .Distinct()
+              .ToList();
+
+            var usedQuizzes = _dbContext.Quizzes
+                .Where(q => quizzes.Contains(q.Id))
+                .Include(q => q.Themes)
+                .ToList();
+
+
+            bool existsInQuizzes = usedQuizzes
+                .Any(q => q.Themes.Any(qt => qt.Id == id));
+
+            if (existsInQuizzes)
+            {
+                return null;
+            }
             Theme theme = _dbContext.Themes.Find(id);
             theme.Name = updatedTheme.Name;
             _dbContext.SaveChanges();
